lxd/db/cluster: Move trigger management to a separate file.
